Encoder: fix input image alloc_size checks
Don't require the last row of an input image to have
a full row stride of data, only require width pixels
of data for the last row.

Usually this isn't a problem, but interlaced JPEG XS
is coded in form of two codestreams, at least in MPEG-TS.
That means that an application might feed an interlaced
image with both fields interleaved row by row to the encoder
in two passes, with a rowstride of double the usual image
stride, so that the encoder only picks up every second row
and encodes each fields individually. This means however
that for the second field, which ends on the last row of
the image, there is only half the passed rowstride of
allocated data available, not a full row stride.

This issue was noticed when implementing support for
interlaced JPEG XS encoding in GStreamer.