CiCd: Fix .pc file for use under Windows (#27)
Include LIB_PKG_CONFIG_EXTRA_CFLAGS as DEC_PKG_CONFIG_EXTRA_CFLAGS
does not exist. On Windows this will expand to "-DDEF_DLL" so that
"dllimport" is enabled in "SvtJpegxs.h".

Rename "EB_DLL" to "DEF_DLL" in Cflags.private. "EB_DLL" does not exist
and by unsetting "DEF_DLL" we make sure that "dllimport" is not used
when linking the static lib.